Challenge Template Re-upload
This is a blank 3D Unity project for students having trouble creating a new file.

This template can be used to start ANY of the prototypes and challenges. When you make changes in Unity (like importing your starter files) you can "Save As" to create a new instance of this project and rename it to the proper assignment. Please reach out via email or Teams if you have any questions.